Redesign hackearth and opensource charts (#13511)
* Refresh /desktop/developers

* Make the linter not sad

* Apply review changes

* Refactor developer-chart.js to match new design

* Refactor hows graphs are drawn

* Apply redesign to hackearth&opensource chars && change some d3 file names for clarity

* Apply design suggestions

* Linty lint

* Use a table inplace of the graph on small screens
