+++ templates/18-04/_release-chart-1604.html
@@ -94,4 +94,4 @@
   </div>
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.2/d3.min.js"></script>
-<script src="{{ versioned_static('js/dist/release-chart.js') }}"></script>
+<script src="{{ versioned_static('js/dist/release-chart-manager.js') }}"></script>
