+++ templates/cpu-compatibility/index.html
@@ -253,6 +253,6 @@
 </section>
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.2/d3.min.js" defer></script>
-<script src="{{ versioned_static('js/dist/release-chart.js') }}" defer></script>
+<script src="{{ versioned_static('js/dist/release-chart-manager.js') }}" defer></script>
 
 {% endblock content %}
