+++ templates/desktop/partial/_developers-table.html
@@ -1,0 +1,46 @@
+<table aria-label="Percentage of operating systems used in organisations to support software infrastructure">
+  <thead>
+    <tr>
+      <th>Operating system</th>
+      <th>Percentage</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>Ubutun</th>
+      <td>35.6%</td>
+    </tr>
+    <tr>
+      <th>Debian</th>
+      <td>21.4%</td>
+    </tr>
+    <tr>
+      <th>CentOS</th>
+      <td>19.5%</td>
+    </tr>
+    <tr>
+      <th>RHEL</th>
+      <td>16%</td>
+    </tr>
+    <tr>
+      <th>OpenSUSE</th>
+      <td>14.6%</td>
+    </tr>
+    <tr>
+      <th>SLES</th>
+      <td>13.5%</td>
+    </tr>
+    <tr>
+      <th>SELinux</th>
+      <td>11.1%</td>
+    </tr>
+    <tr>
+      <th>Rocky Linux</th>
+      <td>7.5%</td>
+    </tr>
+    <tr>
+      <th>NavyLinux</th>
+      <td>4.8%</td>
+    </tr>
+  </tbody>
+</table>
