+++ templates/desktop/partial/_hackearth-table.html
@@ -1,0 +1,50 @@
+<table aria-label="Operating systems developers prefer">
+  <thead>
+    <tr>
+      <th>Operating system</th>
+      <th>Percentage</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>Ubuntu</th>
+      <td>66%</td>
+    </tr>
+    <tr>
+      <th>MS Windows</th>
+      <td>61%</td>
+    </tr>
+    <tr>
+      <th>MacOS</th>
+      <td>57%</td>
+    </tr>
+    <tr>
+      <th>CentOs</th>
+      <td>11%</td>
+    </tr>
+    <tr>
+      <th>Debian</th>
+      <td>10%</td>
+    </tr>
+    <tr>
+      <th>Fedora</th>
+      <td>9%</td>
+    </tr>
+    <tr>
+      <th>Arch Linux</th>
+      <td>8%</td>
+    </tr>
+    <tr>
+      <th>Solaris</th>
+      <td>4%</td>
+    </tr>
+    <tr>
+      <th>FreeBSD</th>
+      <td>2%</td>
+    </tr>
+    <tr>
+      <th>Deepin</th>
+      <td>1%</td>
+    </tr>
+  </tbody>
+</table>
