+++ templates/engage/fr_why-openstack.html
@@ -158,5 +158,5 @@ https://docs.google.com/document/d/1Nvam2qif5UPjTyUPWxhJo1XPiFDM-Eb5FxfSWCbNX8c
 <div class="u-hide" id="contact-form-container" data-form-location="/shared/forms/interactive/openstack" data-form-id="4578" data-lp-id="2086" data-return-url="https://ubuntu.com/openstack/thank-you" data-lp-url="https://pages.ubuntu.com/things-contact-us.html">
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.2/d3.min.js" defer></script>
-<script src="{{ versioned_static('js/dist/release-chart.js') }}" defer></script>
+<script src="{{ versioned_static('js/dist/release-chart-manager.js') }}" defer></script>
 {% endblock content %}
