+++ templates/shared/_kernel-support-schedule.html
@@ -392,5 +392,5 @@
 </section>
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.2/d3.min.js" defer></script>
-<script src="{{ versioned_static('js/dist/release-chart.js') }}" defer></script>
+<script src="{{ versioned_static('js/dist/release-chart-manager.js') }}" defer></script>
 <script src="{{ versioned_static('js/dist/tabbedContent.js') }}"></script>
