+++ templates/shared/_release_schedule.html
@@ -111,4 +111,4 @@
 </div>
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.2/d3.min.js"></script>
-<script src="{{ versioned_static('js/dist/release-chart.js') }}"></script>
+<script src="{{ versioned_static('js/dist/release-chart-manager.js') }}"></script>
